+++ rtiow/V/A08-1.v
@@ -201,9 +201,9 @@ fn main() {
 //				println('rfcnt=$rfcnt riuscnt=$riuscnt riudcnt=$riudcnt')
 			}
 			col = vec.div(col, ns)
-			ir := int(255.99 * col.x)
-			ig := int(255.99 * col.y)
-			ib := int(255.99 * col.z)
+			ir := int(f32(255.99) * col.x)
+			ig := int(f32(255.99) * col.y)
+			ib := int(f32(255.99) * col.z)
 			println('$ir $ig $ib')
 		}
 	}
